Add IsDynamicCodeSupported feature switch support for Mono
