+++ src/mono/System.Private.CoreLib/src/System/Reflection/Emit/AssemblyBuilder.Mono.cs
@@ -206,6 +206,8 @@ namespace System.Reflection.Emit
         [DynamicDependency(nameof(access))] // Automatically keeps all previous fields too due to StructLayout
         private AssemblyBuilder(AssemblyName n, AssemblyBuilderAccess access)
         {
+            EnsureDynamicCodeSupported();
+
             aname = (AssemblyName)n.Clone();
 
             if (!Enum.IsDefined(typeof(AssemblyBuilderAccess), access))
